List 컴포넌트 분류하기
