+++ src/App.js
@@ -1,33 +1,11 @@
 import React, { useState } from "react";
 import "./App.css";
+import Lists from "./components/Lists";
 
 export default function App() {
 
   const [todoData, setTodoData] = useState([]);
   const [value, setValue] = useState("");
-
-
-  const btnStyle = {
-    color: "#fff",
-    border: "none",
-    padding: "5px 9px",
-    borderRadius: "50%",
-    cursor: "pointer",
-    float: "right"
-  }
-
-  const getStyle = (completed) => {
-    return {
-      padding: "10px",
-      borderBottom: "1px #ccc dotted",
-      textDecoration: completed ? "line-through" : "none"
-    }
-  };
-
-  const handleClick = (id) => {
-    let newTodoData = todoData.filter(data => data.id !== id)
-    setTodoData(newTodoData);
-  };
 
   const handleChange = (e) => {
     console.log("e", e.target.value);
@@ -48,16 +26,6 @@ export default function App() {
     setValue("");
   };
 
-  const handleCompletedChange = (id) => {
-    let newTodoData = todoData.map(data => {
-      if(data.id === id) {
-        data.completed = !data.completed;
-      }
-      return data;
-    })
-
-    setTodoData(newTodoData);
-  }
 
     return (
       <div className="container">
@@ -65,13 +33,7 @@ export default function App() {
           <div className="title">
             <h1>할 일 목록</h1>
           </div>
-          {todoData.map((data) => (
-            <div style={getStyle(data.completed)} key={data.id}>
-            <input type="checkbox" defaultChecked={false} onChange={() => handleCompletedChange(data.id)} />
-            {data.title}
-            <button style={btnStyle} onClick={() => handleClick(data.id)}>x</button>
-          </div>
-          ))}
+          <Lists todoData={todoData} setTodoData={setTodoData} />
           <form style={{display: "flex"}} onSubmit={handleSubmit}>
             <input 
               type="text"
